REF refactor the code to finish the tests and get ready for masked coadd sims
