+++ pizza_cutter/slice_utils/interpolate.py
@@ -2,95 +2,6 @@ import numpy as np
 from scipy.interpolate import CloughTocher2DInterpolator
 
 
-def build_slice_locations(*, central_size, buffer_size, image_width):
-    """Build the locations of the slices.
-
-    Parameters
-    ----------
-    central_size : int
-        Size of the central region for metadetection in pixels.
-    buffer_size : int
-        Size of the buffer around the central region in pixels.
-    image_width : int
-        The width of the image in pixels.
-
-    Returns
-    -------
-    row : array-like
-        The row of the slice centers.
-    col : array-like
-        The column of the slice centers.
-    start_row : array-like
-        The starting row of the slice.
-    start_col : array-like
-        The starting column of the slice.
-    """
-    # compute the number of pizza slices
-    # - each slice is central_size + 2 * buffer_size wide
-    #   because a buffer is on each side
-    # - each pizza slice is moved over by central_size to tile the full
-    #  coadd tile
-    # - we leave a buffer on either side so that we have to tile only
-    #  im_width - 2 * buffer_size pixels
-    cen_offset = (central_size - 1) / 2.0  # zero-indexed at pixel centers
-
-    nobj = (image_width - 2 * buffer_size) / central_size
-    if int(nobj) != nobj:
-        raise ValueError(
-            "The coadd tile must be exactly tiled by the pizza slices!")
-    nobj = int(nobj)
-
-    loc = np.arange(nobj) * central_size + buffer_size + cen_offset
-    col, row = np.meshgrid(loc, loc)
-    row = row.ravel()
-    col = col.ravel()
-    start_row = row - buffer_size - cen_offset
-    start_col = col - buffer_size - cen_offset
-
-    return row, col, start_row, start_col
-
-
-def symmetrize_weight(*, weight):
-    """Symmetrize zero weight pixels.
-
-    WARNING: This function operates in-place!
-
-    Parameters
-    ----------
-    weight : array-like
-        The weight map for the slice.
-    """
-    if weight.shape[0] != weight.shape[1]:
-        raise ValueError("Only square images can be symmetrized!")
-
-    weight_rot = np.rot90(weight)
-    msk = weight_rot == 0.0
-    if np.any(msk):
-        weight[msk] = 0.0
-
-
-def symmetrize_bmask(*, bmask, bad_flags):
-    """Symmetrize masked pixels.
-
-    WARNING: This function operates in-place!
-
-    Parameters
-    ----------
-    bmask : array-like
-        The bit mask for the slice.
-    bad_flags : int
-        The flags to symmetrize in the bit mask using
-        `(bmask & bad_flags) != 0`.
-    """
-    if bmask.shape[0] != bmask.shape[1]:
-        raise ValueError("Only square images can be symmetrized!")
-
-    bm_rot = np.rot90(bmask)
-    msk = (bm_rot & bad_flags) != 0
-    if np.any(msk):
-        bmask[msk] |= bm_rot[msk]
-
-
 def _interp_image(*, image, good_msk, bad_msk, yx):
     img_interp = CloughTocher2DInterpolator(
         yx[good_msk, :],
